pybot/pyproject.toml
[tool.poetry]
name = "humblimage"
version = "0.1.0"
description = ""
authors = ["humbldump <humbldump@protonmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.27.1"
python-dotenv = "^0.20.0"
tweepy = "^4.9.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
autopep8 = "^1.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
